bootloader: split handoff security prep into per function modules
detect_cpu_security_features and estimate_tsc_frequency move into their
own files under handoff/prepare/security, leaving the module as an export
map.
