+++ nonos-bootloader/src/handoff/prepare/security/estimate_tsc_frequency.rs
@@ -15,25 +15,10 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::handoff::timing::read_tsc;
-use core::arch::x86_64::{__cpuid, __cpuid_count};
 
-/// Detect SMEP/SMAP/UMIP support. Returns (smep, smap, umip) flags.
-pub fn detect_cpu_security_features() -> (bool, bool, bool) {
-    let max_leaf = __cpuid(0).eax;
-    if max_leaf < 7 {
-        return (false, false, false);
-    }
-    let cpuid7 = __cpuid_count(7, 0);
-    let smep = (cpuid7.ebx & (1 << 7)) != 0;
-    let smap = (cpuid7.ebx & (1 << 20)) != 0;
-    let umip = (cpuid7.ecx & (1 << 2)) != 0;
-    (smep, smap, umip)
-}
-
-/// Estimate TSC frequency by measuring cycles over 10ms stall. Fallback: 2GHz.
 pub fn estimate_tsc_frequency(bs: &uefi::table::boot::BootServices) -> u64 {
     let tsc_start = read_tsc();
-    let _ = bs.stall(10_000); // 10ms in microseconds
+    let _ = bs.stall(10_000);
     let tsc_end = read_tsc();
     if tsc_end > tsc_start {
         (tsc_end - tsc_start) * 100
